Move Conversions.javassistOp to ConvertedOpInfo
